telegram-bot/requirements.txt
aiogram
pyyaml
validators
httpx